Depuración y reorganización de fuentes
