+++ src/components/Buttons/ButtonRightIcon.jsx
@@ -1,7 +1,5 @@
 import React, { useContext } from 'react' 
 import './ButtonRightIcon.css';
-import { Link } from 'react-router-dom';
-import { FaUserCheck } from 'react-icons/fa';
 
 const ButtonRightIcon = ({ title, text, icon }) => {
   return (
