+++ src/components/Footer/Footer.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './footer.css'; 
+import './Footer.css'; 
 import { Link } from 'react-router-dom';
 import { urlLogoFooter } from '../utils/global.context';
