--- src/components/adminPage.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-/*aca es para mostrar la pagina de admin con el container de botnes y el form*/
-
-
-const AdminPage = () => {
-  return (
-    <>
-      <h1>Admin Page</h1>
-      </>
-  );
-};
-
-export default AdminPage;
