+++ src/components/routes/Admin/Caracteristicas/CaracteristicaLista.css
